pyproject.toml
[tool.poetry]
name = "leetcode_prep"
version = "0.1.0"
description = "preparation cli"
authors = ["Ydot19 <abdalhalim.yaadata@gmail.com>"]
packages = [
    { include = "common" }
]


[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.3"
pulumi = "^3.33.2"
pulumi-gcp = "^6.25.0"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.5.0"
black = "22.3.0"
python-dotenv = "^0.20.0"
pandas = "^1.4.2"

[tool.poetry.scripts]
etl-to-db = "etl.main:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
